Started Adding Javadocs to the code. Internal changes to Downloader.
Javadocs are being added to the code to document the methods needed for
implementation. Currently Debugging, Downloader, and
ModuleDescriptionFieldshave been documented.

Downloader has been reworked slightly to allow for the inclusion of
authentication methods. Authentication had previously been the
responsibility of the implementer, but now it is being incorporatated
into the Downloader class itself. Currently only Basic authenticataion
is implemented, but support for OAuth2 authentication is planned. As a
result the openDownloadStream method now has several different variants.
A new constructor has also been added that excludes the fileNum and
numPadding parameters. This is intended for use with files that are not
part of a set.